requirements_pix2pix-zero.txt
accelerate==1.2.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==24.3.0
bleach==6.2.0
blinker==1.9.0
blis==1.1.0
braceexpand==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cmake==3.31.2
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
decorator==5.1.1
decord==0.6.0
diffusers==0.12.1
distlib==0.3.9
einops==0.8.0
executing==2.1.0
fairscale==0.4.4
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.27.0
identify==2.6.4
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
iopath==0.1.10
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lit==18.1.8
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.11
narwhals==1.20.1
networkx==3.4.2
nodeenv==1.9.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
portalocker==3.0.0
pre_commit==4.0.1
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
pyDeprecate==0.3.2
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
salesforce-lavis==1.0.2
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
stack-data==0.6.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
text-unidecode==1.3
thinc==8.3.3
threadpoolctl==3.5.0
tifffile==2024.12.12
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1
torch-fidelity==0.3.0
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.26.1
triton==2.0.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.100
webencodings==0.5.1
wrapt==1.17.0
zipp==3.21.0
